requirements.txt
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.31.2
graphviz==0.19.1
idna==3.3
importlib-metadata==4.11.3
interchange==2021.0.4
ipykernel==6.12.1
ipympl==0.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyterlab==3.3.3
jupyterlab-pygments==0.2.0
jupyterlab-server==2.12.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mchmm==0.4.3
mistune==0.8.4
monotonic==1.6
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
neo4j==4.4.2
nest-asyncio==1.5.5
notebook==6.4.10
notebook-as-pdf==0.5.0
notebook-shim==0.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
pansi==2020.7.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py2neo==2021.2.3
pycparser==2.21
pyee==8.2.2
pygame==2.1.2
Pygments==2.11.2
pyparsing==3.0.7
pypcapfile==0.12.0
PyPDF2==1.27.5
pyppeteer==1.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
read-only-property==0.1
requests==2.27.1
scapy==2.4.5
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tk==0.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
widgetsnbextension==3.6.0
zipp==3.8.0
